+++ app/services/intent/classifier.py
@@ -18,6 +18,7 @@ Pilihan intent:
 
 - reservation
 - view_reservation
+- update_reservation
 - menu
 - promo
 - faq